Extract SessionViewer leaf modules into components/session-viewer/
First slice of the SessionViewer.tsx decomposition (8,018 -> 7,575 lines).
Moves the self-contained module-level pieces into focused files:

- session-viewer/model-labels.ts: prettyModel, metadataModelLabel,
  modelIsCodex, switchDividerText
- session-viewer/model-switches.ts: withModelSwitches
- session-viewer/transcript-cache.ts: CachedTranscriptView + LRU
- session-viewer/transcript-anchor.ts: pickScrollAnchor, holdTranscriptAnchor
- session-viewer/use-live-plan.ts: useLivePlan hook
- session-viewer/busy-indicators.tsx: WorkspaceWaiting, ConversationLoading,
  SteerWaiting, BusyInline (+ private BusyElapsed/BusyStopping)
- session-viewer/streaming-message.tsx: StreamingMessage

Pure code motion, no behavior change. Two call sites read the transcript
LRU map directly; they now go through peekCachedTranscriptView, which is
the same Map.get without the reorder (both sites immediately re-cache).

Typecheck clean; 1,201 frontend tests pass.
